+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Version 0.5.0
+-------------
+
+- Adding lock wrapper for functions
+- Adding duplicate file finder
+- Adding easy CSV / list transformation
+
 Version 0.4.0
 -------------